Refactor: UserDataSource 인터페이스 파일로 분리
